refactor(auth): move refresh token logic to use case and fix typo in service filename
Extract refresh token verification logic from controller to RefreshTokenUserUseCase for better separation of concerns. Also fix typo in token service filename from 'tocken.service' to 'token.service'.
